+++ service/kernel/compression/compression_stream.cpp
@@ -554,7 +554,7 @@ size_t DecompressionStream::getDecompressedDataSize()
         return 0;
     }
 
-    for (int i = 0; i < (*(CBC *)_blocks).size(); i++)
+    for (size_t i = 0; i < (*(CBC *)_blocks).size(); i++)
     {
         decompressBlock(i);
     }
